Struct cudb::db::Collection

pub struct Collection { /* fields omitted */ }

User API for collection-level actions.

Implementations

impl Collection

pub fn insert_one(&self, doc: Document)

Insert one document.

pub fn insert_many(&self, docs: Vec<Document>)

Insert a vector of documents.

pub fn find_one(&self, query: Query) -> Option<Document>

Fetch at most one document matching the query.

pub fn find_many(&self, query: Query) -> Vec<Document>

Fetch a vector of documents matching the query.

pub fn update_one(&self, query: ConstraintDocument, update: UpdateDocument)

Update at most one document that matches the query.

pub fn update_many(&self, query: ConstraintDocument, update: UpdateDocument)

Update all documents matching the query.

pub fn replace_one(&self, query: ConstraintDocument, replace: Document)

Replace at most one document that matches the query.

pub fn delete_one(&self, query: ConstraintDocument)

Delete at most one document that matches the query.

pub fn delete_many(&self, query: ConstraintDocument)

Delete all documents that match the query.

impl Collection

pub fn from(path: String) -> Collection

Create a collection from a path

pub fn create_index(&mut self, ind_names: Vec<FieldPath>)

Create a B-tree index on a set of fields in the collection.

pub fn get_indices(&self) -> &HashMap<Index, BTreeMap<IndexInstance, Offset>>

Get all indices created on this collection.