Struct cudb::mmapv1::Pool

pub struct Pool { /* fields omitted */ }

A linear collection of contiguous blocks.

Implements random-access operations: insertion, update (including resizing/reallocation), (soft) deletion. Also implements a linear scanning. Currently uses a very simple (bad) allocation scheme.

Implementations

impl Pool

pub fn new(path: &Path) -> Pool

Create a new memory pool from a file, creating the file if necessary.

pub fn read_indices(&self) -> Vec<u8>

pub fn write_indices(&mut self, indices_buf: &[u8])

Save indices to file.

pub fn close(self)

Close pool (closes open file).

pub fn drop(self)

Delete pool (deletes file).

pub fn get_size(&self) -> usize

Gets pool (file) size in bytes.

pub fn fetch_block_at_offset(&self, off: Offset) -> TopLevelDocument

Fetch a top level document from a block offset, reading the header.

pub fn fetch(&self, blk: &Block) -> TopLevelDocument

Fetch a top level document from a block address, ignoring the header.

pub fn delete(&mut self, blk: Block)

Delete a block.

pub fn write(&mut self, tldoc: &mut TopLevelDocument)

Update document (including_header), return new document. Note that this may require resizing, which may update the document.

pub fn write_new(&mut self, doc: Document) -> TopLevelDocument

Write a new document (including header), and return the TopLevelDocument.

pub fn scan(&self) -> Vec<TopLevelDocument>

Linearly scan and retrieve all documents from the pool.

pub fn get_next_offset(&self, doc: Document) -> TopLevelDocument