Struct cudb::db::Collection

pub struct Collection { /* fields omitted */ }

User API for collection-level actions.

Implementations

impl Collection

pub fn insert_one(&mut self, doc: Document)

Insert one document.

pub fn insert_many(&mut self, docs: Vec<Document>)

Insert a vector of documents.

pub fn find_one(&mut self, query: Query) -> Option<Document>

Fetch at most one document matching the query.

pub fn find_many(&self, query: Query) -> Vec<Document>

Fetch a vector of documents matching the query.

pub fn find_all(&self) -> Vec<Document>

Fetch all documents from collection.

pub fn update_one(&mut self, query: Query, update: Document)

Update at most one document that matches the query.

pub fn update_many(&mut self, query: ConstraintDocument, update: Document)

Update all documents matching the query.

pub fn delete_one(&mut self, query: Query)

Delete at most one document that matches the query.

pub fn delete_many(&mut self, query: ConstraintDocument)

Delete all documents that match the query.

impl Collection

pub fn from(path: &str) -> Collection

Create a collection from a path.

pub fn get_pool(&self) -> &Pool

Get the collection’s underlying pool.

pub fn get_mut_pool(&mut self) -> &mut Pool

Get the collection’s underlying pool mutably.

pub fn declare_index(&mut self, ind_names: Vec<FieldSpec>)

Create a B-tree index on a list of fields in the collection. If a document doesn’t have a field, will use the default value instead.

pub fn add_document_to_indices(&mut self, top_level_doc: &TopLevelDocument)

Add a document to all existing indices. If any Index is invalid, return without inserting any Index.

pub fn get_indices(
&self
) -> &HashMap<IndexSchema, BTreeMap<Index, HashSet<Offset>>>

Get all indices created on this Collection.

pub fn close(self)

Close collection and underlying file pointers.

pub fn drop(self)

Drop collection.

pub fn get_best_index_schema(
&self,
constraint_schema: &IndexSchema
) -> Option<&IndexSchema>

Get the Index Schema that most closely matches the provided Constraints. Only Index Schemas that fully match the Constraints will be considered.